change package and tuned dagger
