--- app/src/androidTest/java/com/OverSadBoy/samurairise/ExampleInstrumentedTest.kt
@@ -1,21 +1,0 @@
-package com.OverSadBoy.samurairise
-
-import androidx.test.InstrumentationRegistry
-import androidx.test.runner.AndroidJUnit4
-import org.junit.Assert
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() { // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        Assert.assertEquals("com.OverSadBoy.samurairise", appContext.packageName)
-    }
-}
