+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.OverSadBoy.samurairise">
+    package="com.overSadBoy.samurairise">
 
     <application
+        android:name=".dagger.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -11,14 +12,14 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".view.activity.MainActivity">
+        <activity android:name="com.overSadBoy.samurairise.view.activity.alarms.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.activity.ActivityAdd" />
+        <activity android:name="com.overSadBoy.samurairise.view.activity.add.ActivityAdd" />
     </application>
 
 </manifest>
