--- app/src/main/java/com/OverSadBoy/samurairise/dagger/App.kt
@@ -1,19 +1,0 @@
-package com.OverSadBoy.samurairise.dagger
-
-import android.app.Application
-import androidx.test.runner.AndroidJUnit4
-import org.junit.runner.RunWith
-
-class App : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        component = DaggerPresenterComponent.builder().build()
-    }
-
-    companion object {
-        private var component: PresenterComponent? = null
-        fun getComponent(): PresenterComponent? {
-            return component
-        }
-    }
-}
