--- app/src/main/java/com/OverSadBoy/samurairise/dagger/ModelModule.kt
@@ -1,16 +1,0 @@
-package com.OverSadBoy.samurairise.dagger
-
-import androidx.test.runner.AndroidJUnit4
-import com.OverSadBoy.samurairise.model.Model
-import com.OverSadBoy.samurairise.model.database.DataBase
-import dagger.Module
-import dagger.Provides
-import org.junit.runner.RunWith
-
-@Module
-class ModelModule {
-    @Provides
-    fun provideModel(dataBase: DataBase?): Model? {
-        return Model(dataBase)
-    }
-}
