--- app/src/main/java/com/OverSadBoy/samurairise/view/dialog/DialogListener.kt
@@ -1,9 +1,0 @@
-package com.OverSadBoy.samurairise.view.dialog
-
-import androidx.test.runner.AndroidJUnit4
-import org.junit.runner.RunWith
-
-interface DialogListener {
-    open fun positiveClick()
-    open fun negativeClick()
-}
