+++ app/src/main/java/com/overSadBoy/samurairise/dagger/ViewModelFactory.kt
@@ -1,0 +1,24 @@
+package com.overSadBoy.samurairise.dagger
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import javax.inject.Inject
+import javax.inject.Provider
+
+class ViewModelFactory @Inject constructor(
+    private val viewModels: MutableMap<Class<out ViewModel>, Provider<ViewModel>>
+) : ViewModelProvider.Factory {
+
+    init {
+        App.appComponent.inject(this)
+    }
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        val viewModelProvider = viewModels[modelClass]
+            ?: throw IllegalArgumentException("model class $modelClass not found")
+        return viewModelProvider.get() as T
+    }
+
+}
+
