+++ app/src/main/java/com/overSadBoy/samurairise/view/activity/add/AddViewModel.kt
@@ -1,0 +1,5 @@
+package com.overSadBoy.samurairise.view.activity.add
+
+import androidx.lifecycle.ViewModel
+
+class AddViewModel: ViewModel()
