+++ app/src/main/java/com/overSadBoy/samurairise/view/activity/alarms/AlarmsViewModel.kt
@@ -1,0 +1,5 @@
+package com.overSadBoy.samurairise.view.activity.alarms
+
+import androidx.lifecycle.ViewModel
+
+class AlarmsViewModel: ViewModel()
