+++ app/src/main/java/com/overSadBoy/samurairise/view/base/BaseActivity.kt
@@ -1,0 +1,20 @@
+package com.overSadBoy.samurairise.view.base
+
+import android.os.Bundle
+import android.os.PersistableBundle
+import androidx.appcompat.app.AppCompatActivity
+import com.overSadBoy.samurairise.dagger.ViewModelFactory
+import javax.inject.Inject
+
+abstract class BaseActivity : AppCompatActivity() {
+
+    @Inject
+    protected lateinit var viewModelFactory: ViewModelFactory
+
+    protected abstract fun inject()
+
+    override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?) {
+        super.onCreate(savedInstanceState, persistentState)
+        inject()
+    }
+}
