+++ app/src/main/java/com/overSadBoy/samurairise/view/dialog/DialogListener.kt
@@ -1,0 +1,9 @@
+package com.overSadBoy.samurairise.view.dialog
+
+import androidx.test.runner.AndroidJUnit4
+import org.junit.runner.RunWith
+
+interface DialogListener {
+    open fun positiveClick()
+    open fun negativeClick()
+}
