--- app/src/test/java/com/OverSadBoy/samurairise/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package com.OverSadBoy.samurairise
-
-import org.junit.Assert
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        Assert.assertEquals(4, 2 + 2.toLong())
-    }
-}
